rpk: move cmd/version to its own package
And a small typo
