+++ src/go/rpk/pkg/cli/cmd/version/version.go
@@ -9,21 +9,29 @@
 
 package version
 
-import "fmt"
+import (
+	"fmt"
+
+	"github.com/spf13/cobra"
+)
 
 var (
 	version string
 	rev     string
 )
 
-func Version() string {
-	return version
-}
-
-func Rev() string {
-	return rev
-}
-
 func Pretty() string {
-	return fmt.Sprintf("%s (rev %s)", Version(), Rev())
+	return fmt.Sprintf("%s (rev %s)", version, rev)
+}
+
+func NewCommand() *cobra.Command {
+	command := &cobra.Command{
+		Use:   "version",
+		Short: "Check the current version",
+		Long:  "",
+		Run: func(_ *cobra.Command, _ []string) {
+			fmt.Println(Pretty())
+		},
+	}
+	return command
 }
